Sections in api sidebar (#56)
* Add modular classOrder functionality

Rewrites the classOrder formatter to process 1, 1+ or no listed [[classOrder]] sections in the moonwave.toml file. Also shifts the DocSidebar formatting to the plugin rather than LuaClass

* Rewrite of allLuaClass dependent components

These components now take in the formatted allLuaClassNames object. LuaClass can take it in directly, whereas Redirect looks for the first available link name to use as the redirect link

* Clean up typos

* Support classOrder simpler config

Additionally, updated the logic to be more performant and handle mixed cases of sections/no-sections

* Refactor classOrder config parsing

* Fix ClassOrder type

Co-authored-by: Eryn Lynn <eryn@eryn.io>